fix: login 500 error (timezone comparison) + OTP cooldown Redis error
- auth_service.py: Fix 'can't compare offset-naive and offset-aware
  datetimes' in login, email verification, password reset, and
  password change flows. All DB datetime columns are naive (timestamp
  without time zone) so all comparisons and assignments now use
  datetime.now() instead of datetime.now(timezone.utc)
- otp_service.py: Fix 'value is not an integer or out of range' Redis
  error by casting cooldown TTL to int (0.5*60=30.0 → int(30.0)=30)

+++ services/core/service/otp_service.py
@@ -68,7 +68,7 @@ class OTPService:
         redis_client.store_otp(otp_key, otp_code, expires_in=settings.OTP_EXPIRY_MINUTES * 60)
 
         # Set short-lived cooldown key to prevent rapid resends
-        redis_client.set_cache(cooldown_key, "1", ttl=settings.OTP_RESEND_COOLDOWN_MINUTES * 60)
+        redis_client.set_cache(cooldown_key, "1", ttl=int(settings.OTP_RESEND_COOLDOWN_MINUTES * 60))
         
         # Send OTP based on type
         if email and otp_type == "EMAIL":
